pyproject.toml
[project]
name = "foundry-hostedagentframework-demo"
version = "0.1.0"
description = "Agent demo using Microsoft Agent Framework and Microsoft Foundry"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.3",
    "python-dotenv>=1.2.2",
    "agent-framework-core>=1.0.1",
    "agent-framework-openai>=1.0.1",
    "agent-framework-foundry>=1.0.1",
    "agent-framework-foundry-hosting>=1.0.0a260420",
    "agent-framework-orchestrations>=1.0.0b260409",
    "azure-ai-projects>=2.1.0",
    "azure-search-documents>=11.6.0",
    "azure-monitor-opentelemetry>=1.8.7",
    "mcp>=1.27.0",
    "rich>=13.0.0",
    "azure-ai-evaluation[redteam]>=1.16.5",
    "locust>=2.31.0",
    "requests>=2.31.0",
]

[tool.uv]
prerelease = "allow"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
